% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_mosaic.R
\name{bc_mosaic}
\alias{bc_mosaic}
\title{Mosaic plot}
\usage{
bc_mosaic(.data, x.group, y.group, values, fill.values = NULL,
  segment.label = "value", segment.format = NULL,
  total.label = FALSE, total.format = "\%0.0f")
}
\arguments{
\item{.data}{The data frame or tibble containing your data}

\item{x.group}{The column name used to label segments on the x-axis}

\item{y.group}{The column name used to label segments on the y-axis}

\item{values}{The column name showing the value to use to determine the size of the segments}

\item{fill.values}{Fill colors to use for y.group.}

\item{segment.label}{For each segment, show 'value', 'percent', or 'none'? Default is 'value'.}

\item{segment.format}{Format to use to display segment values as a text string. Uses `sprintf()`.}

\item{total.label}{Show the total for each x.group? Default is `FALSE`.}

\item{total.format}{Format to use to display totals as a text string. Uses `sprintf()`.}
}
\description{
Creates a mosaic plot showing rectangular segments sized by both height and width parameters.
}
